Make program.c set ARGS and PROGRAM_FILE via argv.
The repl sets ARGS and PROGRAM_FILE via code in Base._start(), which
isn't called by compiled programs. This commit adds logic to program.c
to set Core.ARGS, Base.ARGS (the global ARGS), and PROGRAM_FILE, all in
exactly the same way as the REPL. It then passes a reference to
Base.ARGS as the argument to `julia_main(ARGS)`.

This replicates the behavior from the repl which sets:
  PROGRAM_FILE = argv[0]
  Base.ARGS = argv[1:argc]
  Core.ARGS = argv[0:argc]