remove `Base.@ccallable` from doc examples
After upgrading from julia 1.5 to 1.6 my app was no longer compiling, with the error:

```
ERROR: @ccallable was already defined for this method name
```

In the docs I then read that this annotation is no longer needed, however I still saw it in use in the docs. Hence I now remove the annotation from the docs.