Fix lazy artifact evaluation
The string literal is evaluated at code loading time so the
`i686-w64-mingw32` is required on both 32 and 64 bit platforms.
The artifact string also checks for platform so on 64 bit platform
the 32 bit version will be filtered out.

Protect them in @static blocks so only the correct ones are used
for each platform.