Allow the user to define arbitrary C variables.
Since PC.jl now exposes `c_driver_program` which allows the users to
specify their own C wrapper program, it would be nice to be able to pass
in more variables which might be best defined at build time.

For my use case, I would like to bake in a build version at build time.
And rather than storing it in julia or needing to compile it into a
file, it is far easeier to just have something like:

  #ifdef VERSION
              printf("%s\n", VERSION);
  #else

This change will allow for that by passing a `c_defines` dict and
letting the user pass in arguments for the compiler to define as a
macro.