run `get_compiler_cmd` earlier, to error if not found
Closes: #677, #714, #716, #727

In many scenraios a default user shell may not have access to a compiler.
The current behavior runs the checks after compiling ther sysimg, which may take many
minutes. This moves the check earlier so that this error happens earlier.
It also splits the `cmd` and `args` such that it may be re-used or
potentially altered by a user.