Revert "Allow setting a command prefix for the sysimage build command"
It is not necessary to provide a prefix in the build phase but it may be
necessary when the script given by `precompile_execution_file` is executed,
e.g. if some package requires the script to be run in an MPI context
via `mpirun`.
Thus, this reverts commit d3b7edccf535e70452c28607dad7f9f686b92b8c.