Modify MyLib to demonstrate bad pointer return value
This is an MWE that demonstrates that a `Ptr{Cvoid}` can end up mangled
in certain cases when returned from a `@ccallable`.

I don't know why this is happening, but it does seem quite fragile so
it's probably not a simple ABI definition issue. For example, printing
the value out on the Julia side is enough to make the problem go away.